requirements.txt
# Core
python>=3.9
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
